refactor: 중복 파일 제거 (#4)
