+++ src/main/java/com/newfit/reservation/dto/request/CreateProposalRequestDto.java
@@ -1,7 +1,6 @@
 package com.newfit.reservation.dto.request;
 
-import com.newfit.reservation.domain.Proposal;
-import com.newfit.reservation.domain.Report;
+import com.newfit.reservation.domain.dev.Proposal;
 import lombok.Getter;
 import lombok.Setter;
